requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
